Big commit
Turn the bookmarklet into a Chrome extension so that we can
automatically set all of the links to open in a new tab, without
having to click on bookmarklet each time we come to the website.
Also, since this is now a Chrome extension, we can leverage the
Chrome extension API to allow for opening links in new tabs
without changing focus to the new tab.